+++ doc/source/doxygen-docs/changelog.md
@@ -7,12 +7,15 @@
   - \ref mrpt_gui_grp
     - Useless nanogui_win() converted into mrpt::gui::CDisplayWindowGUI::nanogui_screen()
     - nanogui: New methods: nanogui::Screen::mouseState(), nanogui::Screen::mouseModifiers()
+  - \ref mrpt_opengl_grp
+    - New load flags in mrpt::opengl::CAssimpModel::loadScene()
 - BUG FIXES:
   - RawLogViewer: Fix wrong indices in tree view. Fix freezed progress bar loading a second rawlog.
   - RawLogViewer: Fix wrong rendering if font is missing in the system (Ubuntu 20.04)
   - rawlog-edit: Fix --cut operation leaving empty sensory frames/action collections.
   - mrpt::opengl::CCylinder::setHasBases() was ignored since last OpenGL3 refactor.
   - Fix building against OpenCV 4.4
+  - Correct texture loading in mrpt::opengl::CAssimpModel.
 
 ------
 # Version 2.1.0: Released Aug 31st, 2020